requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
decorator==4.4.2
future==0.18.2
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
python-dotenv==0.15.0
python-telegram-bot==13.2
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
